🎨 encapsulated reward functions in a separate file
